Image data source folder restructured, image repository created it will manage image data
