+++ app/src/main/java/com/gokhanaliccii/infiniteimagelisting/datasource/image/ImageDataSource.kt
@@ -1,0 +1,10 @@
+package com.gokhanaliccii.infiniteimagelisting.datasource.image
+
+interface ImageDataSource {
+
+    fun loadImages(count: Int = 20, page: Int = 1, loadCallBack: ImageLoadCallBack)
+
+    interface ImageLoadCallBack {
+        fun onImagesLoaded(images: List<ImageUIModel>)
+    }
+}
