+++ app/src/main/java/com/gokhanaliccii/infiniteimagelisting/datasource/image/ImageRepository.kt
@@ -1,0 +1,6 @@
+package com.gokhanaliccii.infiniteimagelisting.datasource.image
+
+class ImageRepository {
+
+
+}
