+++ app/src/main/java/com/gokhanaliccii/infiniteimagelisting/datasource/image/ImageUIModel.kt
@@ -1,4 +1,4 @@
 package com.gokhanaliccii.infiniteimagelisting.datasource.image
 
-class ImageResponse {
+class ImageUIModel {
 }
